+++ test/controllers/designs_controller_test.rb
@@ -724,6 +724,27 @@ class DesignsControllerTest < ActionController::TestCase
     assert_template 'update'
   end
 
+  test "should update an existing variable on a design" do
+    put :update, id: designs(:sections_and_variables), project_id: @project, variable: { name: "var_date_updated", display_name: "Today's Date Updated", branching_logic: '1 = 1' }, position: 0, variable_id: variables(:date).id, update: 'variable', format: 'js'
+    assert_not_nil assigns(:project)
+    assert_not_nil assigns(:design)
+    assert_equal 11, assigns(:design).options.size
+    assert_equal 'var_date_updated', assigns(:design).variable_at(0).name
+    assert_equal "Today's Date Updated", assigns(:design).variable_at(0).display_name
+    assert_equal '1 = 1', assigns(:design).options[0][:branching_logic]
+    assert_template 'update'
+  end
+
+  test "should not update an existing variable with a blank name on a design" do
+    put :update, id: designs(:sections_and_variables), project_id: @project, variable: { name: "" }, position: 0, variable_id: variables(:date).id, update: 'variable', format: 'js'
+    assert_not_nil assigns(:project)
+    assert_not_nil assigns(:design)
+    assert_equal 11, assigns(:design).options.size
+    assert_equal 'var_date', assigns(:design).variable_at(0).name
+    assert_equal 1, assigns(:errors).size
+    assert_template 'update'
+  end
+
   test "should remove a variable from a design" do
     put :update, id: @design, project_id: @project, position: 0, delete: 'variable', format: 'js'
     assert_not_nil assigns(:project)
@@ -740,6 +761,25 @@ class DesignsControllerTest < ActionController::TestCase
     assert_not_nil assigns(:design)
     assert_equal 3, assigns(:design).options.size
     assert_equal 'new_domain_for_variable', assigns(:design).variable_at(0).domain.name
+    assert_template 'update'
+  end
+
+  test "should update an existing domain on a design" do
+    put :update, id: designs(:sections_and_variables), project_id: @project, domain: { name: 'dropdown_options_new', display_name: 'New Domain For Dropdown Variable', option_tokens: [ { option_index: 'new', name: 'Easy', value: '1' }, { option_index: 'new', name: 'Medium', value: '2' }, { option_index: 'new', name: 'Hard', value: '3' }, { option_index: 'new', name: 'Old Value', value: 'Value' } ] }, position: 3, variable_id: variables(:dropdown).id, update: 'domain', format: 'js'
+    assert_not_nil assigns(:project)
+    assert_not_nil assigns(:design)
+    assert_equal 11, assigns(:design).options.size
+    assert_equal 'dropdown_options_new', assigns(:design).variable_at(2).domain.name
+    assert_template 'update'
+  end
+
+  test "should not update an existing domain with blank name on a design" do
+    put :update, id: designs(:sections_and_variables), project_id: @project, domain: { name: '', display_name: 'New Domain For Dropdown Variable', option_tokens: [ { option_index: 'new', name: 'Easy', value: '1' }, { option_index: 'new', name: 'Medium', value: '2' }, { option_index: 'new', name: 'Hard', value: '3' }, { option_index: 'new', name: 'Old Value', value: 'Value' } ] }, position: 3, variable_id: variables(:dropdown).id, update: 'domain', format: 'js'
+    assert_not_nil assigns(:project)
+    assert_not_nil assigns(:design)
+    assert_equal 11, assigns(:design).options.size
+    assert_equal 'dropdown_options', assigns(:design).variable_at(2).domain.name
+    assert_equal 1, assigns(:errors).size
     assert_template 'update'
   end